Add files via upload
An early project using data collected from a mildly well-known coffee / beverage chain.   As I was training for analytical work, the main goal here was EDA with less priority on the end model, LightGBM with PCA.  You will note that parameter tuning for LightGBM was not used, this is because I was a noob.